Implemented clear chat feature

+++ src/main/java/me/lightningreflex/lightningutils/configurations/impl/LangConfig.java
@@ -125,7 +125,7 @@ public class LangConfig extends Config {
         }
     }
 
-    public float lang_version;
+    public double lang_version;
     //------------------------------------------------
 
     public LangConfig load(final String path) {
